Move compareWavFiles to helper class as per inputs
