+++ host/tests/rmfAudio_L3_TestCases/rmfAudio_test05_combinedDataCapture.py
@@ -60,13 +60,8 @@ class rmfAudio_test05_combinedDataCapture(rmfAudioHelperClass):
         Returns:
             bool: True if audio capture is a match, False otherwise.
         """
-        test = self.testSetup.get("assets").get("device").get(self.testName)
-
-        # Download test streams to device
-        url =  test.get("streams")
-        if url[index] is not None:
-            ref_file_path = self.testStreams[index]
-            return self.testrmfAudio.compareWavFiles(ref_file_path, file_path)
+        ref_file_path = self.testStreams[index]
+        return self.compareWavFiles(ref_file_path, file_path)
 
     def testFunction(self):
         """
